requirements.txt
musco-pytorch==1.0.6
numpy==1.16.1
tensorly==0.4.5
flopco-pytorch==0.1.4
torch==1.11.0+cu113
torchfile==0.1.0
torchvision==0.12.0+cu113
